requirements.txt
bidict~=0.22.1
frozendict~=2.3.8
orjson~=3.9.2
python-rapidjson~=1.10
regex>=2023.3.23

empire_commons@https://github.com/Tombmyst-Empire/empire-commons/archive/refs/heads/master.zip